pyproject.toml
[tool.poetry]
name = "aws-cdk-elasticache-experiment"
version = "0.0.1"
description = "AWS CDK deployment of Elasticache experiment with Redis on AWS"
authors = ["Santiago Garcia Arango <san99tiago@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "2.115.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.3"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
coverage = "^7.3.2"
boto3 = "^1.29.3"


[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = [
    "cdk",
    "src",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
deploy = "cdk deploy --require-approval never"
test-unit = "coverage run -m pytest tests/unit"
test-unit-html = "coverage html"

[tool.coverage.run]
branch = true
source = ["src", "cdk"]
omit = [
    "**/__init__.py"
]

[tool.coverage.report]
show_missing = false
